requirements.txt
streamlit
pandas
plotly
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl